Revise handling of view data
We extract() these instead of making them available via magic methods.
That's just as good as before; we still need to make sure not to echo
variables directly.  The bonus: we can get PHPStan analysis based on
assumptions of the types, which also serves as documentation.

While this is not foolproof, it seems to be better than no analysis at
all.